+++ src/main/java/flab/project/data/dto/model/HashTag.java
@@ -1,4 +1,4 @@
-package flab.project.data.dto.common;
+package flab.project.data.dto.model;
 
 import io.swagger.v3.oas.annotations.media.Schema;
