+++ src/main/java/flab/project/data/dto/request/UpdateOptionsDto.java
@@ -1,6 +1,6 @@
 package flab.project.data.dto.request;
 
-import flab.project.data.dto.common.Options;
+import flab.project.data.dto.model.Options;
 
 public class UpdateOptionsDto extends Options {
