+++ src/main/java/flab/project/data/dto/request/UpdateProfileDto.java
@@ -1,6 +1,6 @@
 package flab.project.data.dto.request;
 
-import flab.project.data.dto.common.UpdatePage;
+import flab.project.data.dto.model.UpdatePage;
 
 public class UpdateProfileDto extends UpdatePage {
