+++ src/main/java/flab/project/data/dto/response/GetOptionsDto.java
@@ -1,6 +1,6 @@
 package flab.project.data.dto.response;
 
-import flab.project.data.dto.common.Options;
+import flab.project.data.dto.model.Options;
 
 public class GetOptionsDto extends Options {
