+++ src/main/java/flab/project/data/dto/response/GetProfileUpdatePageDto.java
@@ -1,10 +1,8 @@
 package flab.project.data.dto.response;
 
-import flab.project.data.dto.common.UpdatePage;
+import flab.project.data.dto.model.UpdatePage;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
-
-import java.util.List;
 
 @Getter
 @Schema(description = "프로필 페이지 요청때 반환되는 Schema")
